Refactor: moving duplicated code to only one function
