+++ src/pages/Home.mkpe1.js
@@ -5,17 +5,9 @@ import { currentMember } from 'wix-members';
 
 import {
     utils_set_sections_history,
-    utils_config_items,
+    utils_load_history,
     SECTION_STATE_LOADING,
-    SECTION_STATE_DATA,
-    SECTION_STATE_NO_DATA,
 } from 'public/utils';
-
-import {
-    be_mod_utils_get_history,
-} from 'backend/be_mod_utils';
-
-const QTDE_ITENS_RESUMO = 2; // max number of transactions on resume
 
 
 async function fill_member_picture() {
@@ -25,23 +17,9 @@ async function fill_member_picture() {
 	$w("#textOlaNome").text = `Olá, ${member.contactDetails.firstName}`;
 }
 
-async function load_resumed_history() {
-    $w("#repeaterHist").onItemReady( ($item, itemData, index) => {
-        utils_config_items($item, itemData)
-        utils_set_sections_history(SECTION_STATE_DATA);
-    });
-    
-    let history = (await be_mod_utils_get_history()).reverse();
-    history = history.length >= QTDE_ITENS_RESUMO ? history.slice(0 , QTDE_ITENS_RESUMO) : history; // limits items on resume
-    $w("#repeaterHist").data = history; 
-    
-    if (!history.length)
-        utils_set_sections_history(SECTION_STATE_NO_DATA);
-}
-
 $w.onReady(function () {
     utils_set_sections_history(SECTION_STATE_LOADING);
-    load_resumed_history();
+    utils_load_history(true);
 	fill_member_picture();
 
     // Escreva seu código JavaScript aqui usando o API de framework do Velo
