+++ wix.config.json
@@ -1,4 +1,4 @@
 {
   "siteId": "48a7ceaf-d142-4928-9249-98edd3ce37ef",
-  "uiVersion": "191"
+  "uiVersion": "194"
 }
